backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
tensorflow==2.18.0
Pillow==11.1.0
numpy==1.26.4
